pyproject.toml
[tool.poetry]
name = "faq-generator"
version = "0.1.0"
description = "a faq generator"
authors = ["Stefan Hartmann <stefan.hartmann@hartie-labs.com>"]
license = "private"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "3.11.9"
alembic = "^1.13.2"
openai = "^1.50.1"
greenlet = "^3.0.3"
fastapi = "^0.112.2"
pydantic = {extras = ["email"], version = "^2.8.2"}
pydantic-settings = "^2.4.0"
SQLAlchemy = "^2.0.32"
pgvector = "^0.3.2"
uvicorn = "^0.30.6"
PyGithub = "2.4.0"
requests = "^2.32.3"
beautifulsoup4 = "^4.12.3"
stripe = "^10.9.0"
asyncpg = "^0.29.0"
langchain = "^0.3.1"
langgraph = "0.2.28"
langchain-openai ="0.2.1"
pre-commit = "4.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
